Restored the DBF table import (signature of implemented function did not match the one declared in parent class)
git-svn-id: https://svn.52north.org/svn/ilwis/drawers@2735 b64c5c0c-1e74-440b-9268-4ad09cf26390

+++ Sources/Engine/DataExchange/TableExternalFormat.h
@@ -98,7 +98,7 @@ class TableExternalFormat : public TableStore
 		virtual bool    SplitLine(CStdioFile& sIf, FieldValues& result, long& iRec) { return false;};
 		virtual void    AddToTableDomain(FieldValues &line, int iKeyColumn);
 		virtual void    Load();
-		virtual void    PutExternalField(FieldValues &line, int &iMemoCol, int iRec, int iField) {}
+		virtual void    PutExternalField(FieldValues &line, int &iMemoCol, long iRec, int iField) {}
 		virtual long    iNewRecord(long iRec);
 
 		int             iGetKeyColumnIndex();
